fixed type conversion warning
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@2930 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Utilities/IOUtils.h
@@ -313,7 +313,7 @@ void WritePathLinkConfigurations(const string _outputFile,
     vector<CFG>& _path, 
     Environment* _env){
   vector<Cfg*> ppath;
-  for(int i=0; i<_path.size(); i++)
+  for(size_t i=0; i<_path.size(); i++)
     ppath.push_back(&_path[i]);
   WritePathLinkConfigurations(_outputFile, ppath, _env);
 }
